pyproject.toml
[project]
name = "acere"
version = "1.0.0b4"
description = ""
requires-python = ">=3.14,<4.0"
dependencies = [
    # From Template
    "fastapi[standard-no-fastapi-cloud-cli]>=0.118.0",
    "sqlmodel<1.0.0,>=0.0.25",
    "pydantic>=2.11",
    "pydantic-settings>=2.11",
    "pyjwt<3.0.0,>=2.10.1",
    "uvicorn>=0.37.0",
    # Me
    "beautifulsoup4>=4.14.2",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "psutil>=7.1.0",
    "pytz>=2025.2",
    "aiohttp[speedups]>=3.13.0", # Requires python 3.14+ for zstd support
    "rich>=14.1.0",
    "argon2-cffi>=25.1.0",
    "uvloop>=0.22.1",
    "starlette-compress>=1.6.1",
]

[project.optional-dependencies]
type = [
    "mypy",
    "ty",
    "types-passlib>=1.7.7",
    "lxml-stubs>=0.5.1",
    "types-psutil>=7.0.0.20250601",
    "types-pytz>=2025.2.0.20250516",
]
lint = ["ruff"]
test = [
    "pytest >=9.0",
    "pytest-cov",
    "pytest-mock",
    "pytest-random-order",
    "detect-test-pollution",
    "pytest-asyncio",
    "pytest-socket>=0.7.0",
]


[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["acere"]
module-root = ""

[project.scripts]
acerestreamer-scrape = "acere.cli.scraper:__main__.main"

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # From the template
    "B904", # Allow raising exceptions without from e, for HTTPException

    # Specific rules
    "TD002",  # Verbose TODO, only I work on this project
    "TD003",  # Verbose TODO, only I work on this project
    "ISC001", # conflict in the formatter
    "COM812", # conflict in the formatter
    "D103",   # Missing docstring in public function
    "N805",   # field_validator needs cls
    "TRY400", # Fine to log exceptions as errors
    "SIM117", # Double 'with' is more readable with async
    "D1",     # Not going to document everything, we have good type checking
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "S101",    # Allow use of assert in tests
    "ARG001",  # Allow unused arguments in tests (fixtures)
    "E402",    # I do some dumb imports in tests
    "TRY",     # Exceptions in tests are allowed to be messy
    "D",       # Docstrings in tests are optional
    "S",       # No real crypto in tests
    "SLF001",  # Can access private members in tests
    "PLR2004", # Allow magic comparisons in tests
    "ANN401",  # *args: Any, **kwargs: Any in tests
]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
    "pydantic.BaseModel",
    "pydantic_settings.BaseSettings",
    "sqlmodel.SQLModel",
    "uuid.UUID",
    "acere.api.deps.SessionDep",
] # These aren't for type checking only

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 40


[tool.coverage.run]
source = ["acere"]
context = "test"


[tool.coverage.report]
show_missing = false
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.pytest.ini_options]
addopts = [
    "--cov=acere",
    "--cov-report=term",
    "--cov-report=html",
    "--disable-socket",
    "--allow-hosts=127.0.0.1,127.0.1.1",
    "--asyncio-mode=auto",
]


[tool.mypy]
strict = true
files = ["acere"]
exclude = ["venv", ".venv"]
show_error_codes = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
    "type-arg",
    "unused-awaitable",
    "possibly-undefined",
    "redundant-self",
]

[[tool.ty.overrides]]
# Not sure what ty is upset about here
include = ["acere/main.py"]
[tool.ty.overrides.rules]
"invalid-argument-type" = "ignore"
